refactor(runtime): isolate connection cache and transport ownership
