--- docs/refactor/dedupe-2026-08-01.md
@@ -1,23 +1,0 @@
----
-summary: 'Tracks the four-commit internal deduplication and ownership refactor.'
-read_when:
-  - 'Continuing the August 2026 internal refactor work order'
----
-
-# Internal deduplication refactor
-
-## Commit 1: pure utilities and dead internals
-
-Completed. The shared foundations now own stable JSON serialization, daemon config-layer normalization and collection, daemon delay/process liveness, generic token consumption, timeout racing, and child-exit waiting. Generated-command parsing and tool-selection validation now use their canonical helpers, direct runtime types replace dynamic type expressions, and every audit-listed dead internal was re-verified before removal.
-
-## Commit 2: shared CLI contracts
-
-Completed. Record and replay now share config writing, subprocess propagation, and environment instructions. Call, list, and auth render one structured ad-hoc help definition with their original alignment. The two intentionally distinct definition serializers share only their common field projection, and Chrome's standalone fallback is rendered from the canonical patch implementation.
-
-## Commit 3: OAuth persistence and refresh policy
-
-Completed. The public OAuth persistence facade now delegates to internal directory, vault, and composite stores, while expiry classification and both OAuth and refreshable-bearer refresh policy live in a dedicated module. Store snapshots reconcile identity once for read-many migration, and client-secret lookup is canonical without changing the stricter refresh-path handling of blank environment values.
-
-## Remaining commit
-
-Isolate runtime connection cache and transport ownership, then delete this tracker.
